pyproject.toml
[project]
name = 'phi-agents'
version = "0.1.0"
requires-python = ">= 3.10"

[tool.poetry]
name = "phi_agents"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
python = "~3.12"
vllm = "~0.10.2"
torchtune = "~0.4.0"
torchao = "^0.10.0"
ray = "^2.49.2"
aioboto3 = "^13.4.0"
numpy = "^1.0.0"
tqdm = "^4.67.1"
argparse-dataclass = "^2.0.0"
bokeh = "^3.6.3"
cairosvg = "^2.7.1"
cattrs = "^24.1.2"
compress-pickle = { extras = ["lz4"], version = "^2.1.0" }
huggingface-hub = "<1.0"
hydra-core = "^1.3.2"
joblib = "^1.4.2"
matplotlib = "^3.10.0"
mock = "^5.1.0"
names = "^0.3.0"
openai = "^1.63.2"
packaging = "^24.2"
pathvalidate = "^3.2.3"
plotly = "^6.0.0"
psutil = "^7.0.0"
pytest = "^8.3.4"
pyyaml = "^6.0.2"
pyparsing = "^3.2.1"
regex = "^2024.11.6"
requests = ">=2.32.4"
setuptools = ">=78,<80"
tenacity = "^9.0.0"
termcolor = "^2.5.0"
types-python-dateutil = "^2.9.0.20241206"
types-pytz = "^2025.1.0.20250204"
types-requests = "^2.32.0.20241016"
wandb = "^0.19.11"
wandb-workspaces = "^0.1.12"
deepdiff = "^8.2.0"
dill = "0.3.8"
docstring-parser = "^0.16"
fire = "^0.7.0"
hf-transfer = "^0.1.9"
ijson = "^3.3.0"
virtualenv = ">=20.29.2,<20.33.1"
ipython = "^8.32.0"
pre-commit = "^4.1.0"
ipykernel = "^6.29.5"
nbclient = "^0.10.2"
nbformat = "^5.10.4"
ruff = "0.5.5"
jupytext = "^1.16.7"
rich = "^13.9.4"
datasets = "^3.3.2"
munch = "^4.0.0"
litellm = "^1.63.14"
cut-cross-entropy = { git = "https://github.com/apple/ml-cross-entropy.git" }
transformers = "~4.56.1"
accelerate = "~1.6.0"
peft = "^0.15.2"
torch = "~2.8"
torchvision = "^0.23"
nvidia-ml-py = "^12.575.51"
pillow = ">=11.3.0,<12"
protobuf = ">=4.25.8"
starlette = ">=0.47.2"
urllib3 = ">=2.5.0"
aiohttp = ">=3.12.14"

[tool.setuptools]
packages = ["phi_agents"]

[tool.ruff]
line-length = 100
include = ["phi_agents/**.py", "scripts/**.py", "tests/**.py", "web_api/**.py"]
target-version = "py310"

[tool.ruff.lint]
select = ["B", "D", "E", "F", "W", "FLY", "I", "UP", "TCH", "PTH", "D417"]
ignore = [
    "B017",
    "E501",
    "E203",
    "E722",
    "E402",
    "F541",
    "E741",
    "PTH123",
    "D107",
    "D105",
    "D104",
    "D103",
    "D102",
    "D101",
    "D100",
    "D205",   # annoying warnings about docstring formatting
    "D212",   # same
]
exclude = ["**/notebooks/**.py"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D"]
"web_api/*" = ["D"]
"notebooks/*" = ["B018"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-third-party = ["wandb"]

[tool.pytest.ini_options]
addopts = "-s -rA --durations=10"

filterwarnings = [
    "error",
    # note the use of single quote below to denote "raw" strings in TOML
    'ignore::DeprecationWarning',
    'ignore:distutils Version classes are deprecated.*',
    'ignore::SyntaxWarning',
    'ignore:Field "model_id":UserWarning',
]

markers = ["notebook: test all jupyter notebooks and make sure they run"]


